+++ CHANGELOG.md
@@ -25,3 +25,4 @@
 ## [0.4.0]
 
 - Added `Cancel running` button on StatusBar to cancel current running operation.
+- Adeed notebooks support for running go code from the `*.gobook` files.